requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.8
django-cors-headers==3.1.1
django-filter==2.2.0
django-crispy-forms==1.8.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
idna==2.8
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6